+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/Epills.iml" filepath="$PROJECT_DIR$/Epills.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/ePills.iml" filepath="$PROJECT_DIR$/ePills.iml" />
       <module fileurl="file://$PROJECT_DIR$/libs/library-clock/library-clock.iml" filepath="$PROJECT_DIR$/libs/library-clock/library-clock.iml" />
     </modules>
   </component>
